Updates to srpe, including some refactoring
